requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI API (with Vision support)
openai==1.10.0

# Image Processing (for validation)
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
requests==2.31.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Background Tasks (optional)
celery==5.3.6
redis==5.0.1

email-validator
